+++ src/main/java/org/bytedeco/javacpp/Loader.java
@@ -31,6 +31,7 @@ import java.io.OutputStream;
 import java.net.URL;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -108,22 +109,25 @@ public class Loader {
     /**
      * Loads the {@link Properties} associated with the default {@link #getPlatform()}.
      *
-     * @see #loadProperties(String)
+     * @return {@code loadProperties(getPlatform(), null)}
+     * @see #loadProperties(String, String)
      */
     public static Properties loadProperties() {
         String name = getPlatform();
         if (platformProperties != null && name.equals(platformProperties.getProperty("platform"))) {
             return platformProperties;
         }
-        return platformProperties = loadProperties(name);
+        return platformProperties = loadProperties(name, null);
     }
     /**
      * Loads from resources the default {@link Properties} of the specified platform name.
+     * The resource must be at {@code "org/bytedeco/javacpp/properties/" + name + ".properties"}.
      *
      * @param name the platform name
+     * @param defaults the fallback platform name (null == "generic")
      * @return the Properties from resources
      */
-    public static Properties loadProperties(String name) {
+    public static Properties loadProperties(String name, String defaults) {
         Properties p = new Properties();
         p.put("platform", name);
         p.put("platform.path.separator", File.pathSeparator);
@@ -140,7 +144,7 @@ public class Loader {
                 p.load(is);
             }
         } catch (Exception e) {
-            name = "properties/generic.properties";
+            name = "properties/" + defaults + ".properties";
             InputStream is2 = Loader.class.getResourceAsStream(name);
             try {
                 try {
@@ -400,33 +404,47 @@ public class Loader {
         return s.equals("true") || s.equals("t") || s.equals("");
     }
 
+    /** Returns {@code load(getCallerClass(2), loadProperties(), false)}. */
+    public static String load() {
+        return load(getCallerClass(2), loadProperties(), false);
+    }
     /**
      * Loads native libraries associated with the {@link Class} of the caller.
-     * @return {@code load(getCallerClass(2)) }
+     *
+     * @param pathsFirst search the paths first before bundled resources
+     * @return {@code load(getCallerClass(2), loadProperties(), pathsFirst) }
      * @see #getCallerClass(int)
-     * @see #load(Class)
+     * @see #load(Class, Properties, boolean)
      */
-    public static String load() {
+    public static String load(boolean pathsFirst) {
         Class cls = getCallerClass(2);
-        return load(cls);
+        return load(cls, loadProperties(), pathsFirst);
+    }
+    /** Returns {@code load(cls, loadProperties(), false)}. */
+    public static String load(Class cls) {
+        return load(cls, loadProperties(), false);
     }
     /**
      * Loads native libraries associated with the given {@link Class}.
      *
      * @param cls the Class to get native library information from
+     * @param properties the platform Properties to inherit
+     * @param pathsFirst search the paths first before bundled resources
      * @return the full path to the main file loaded, or the library name if unknown
-     *         (but {@code if (!loadLibraries || cls == null) { return null; }})
+     *         (but {@code if (!isLoadLibraries() || cls == null) { return null; }})
      * @throws NoClassDefFoundError on Class initialization failure
      * @throws UnsatisfiedLinkError on native library loading failure
+     * @see #findLibrary(Class, ClassProperties, String, boolean)
+     * @see #loadLibrary(URL[], String)
      */
-    public static String load(Class cls) {
+    public static String load(Class cls, Properties properties, boolean pathsFirst) {
         if (!isLoadLibraries() || cls == null) {
             return null;
         }
 
         // Find the top enclosing class, to match the library filename
         cls = getEnclosingClass(cls);
-        ClassProperties p = loadProperties(cls, loadProperties(), true);
+        ClassProperties p = loadProperties(cls, properties, true);
 
         // Force initialization of all the target classes in case they need it
         List<String> targets = p.get("target");
@@ -461,7 +479,7 @@ public class Loader {
         UnsatisfiedLinkError preloadError = null;
         for (String preload : preloads) {
             try {
-                URL[] urls = findLibrary(cls, p, preload);
+                URL[] urls = findLibrary(cls, p, preload, pathsFirst);
                 loadLibrary(urls, preload);
             } catch (UnsatisfiedLinkError e) {
                 preloadError = e;
@@ -470,7 +488,7 @@ public class Loader {
 
         try {
             String library = p.getProperty("platform.library");
-            URL[] urls = findLibrary(cls, p, library);
+            URL[] urls = findLibrary(cls, p, library, pathsFirst);
             return loadLibrary(urls, library);
         } catch (UnsatisfiedLinkError e) {
             if (preloadError != null && e.getCause() == null) {
@@ -481,16 +499,19 @@ public class Loader {
     }
 
     /**
-     * Finds where the library may be extracted and loaded among the {@link Class}
-     * resources. But in case that fails, also searches the paths found in the
-     * "platform.preloadpath" and "platform.linkpath" properties.
+     * Finds from where the library may be extracted and loaded among the {@link Class}
+     * resources. But in case that fails, and depending on the value of {@code pathsFirst},
+     * either as a fallback or in priority over bundled resources, also searches the paths
+     * found in the "platform.preloadpath" and "platform.linkpath" class properties as well as
+     * the "java.library.path" system property, in that order.
      *
      * @param cls the Class whose package name and {@link ClassLoader} are used to extract from resources
      * @param properties contains the directories to scan for if we fail to extract the library from resources
      * @param libnameversion the name of the library + "@" + optional version tag
+     * @param pathsFirst search the paths first before bundled resources
      * @return URLs that point to potential locations of the library
      */
-    public static URL[] findLibrary(Class cls, ClassProperties properties, String libnameversion) {
+    public static URL[] findLibrary(Class cls, ClassProperties properties, String libnameversion, boolean pathsFirst) {
         String[] s = libnameversion.split("@");
         String libname = s[0];
         String version = s.length > 1 ? s[s.length-1] : "";
@@ -524,34 +545,36 @@ public class Loader {
             }
         }
 
-        int k = 0;
         List<String> paths = new ArrayList<String>();
         paths.addAll(properties.get("platform.preloadpath"));
         paths.addAll(properties.get("platform.linkpath"));
-        URL[] urls = new URL[styles.length * (1 + paths.size())];
+        String libpath = System.getProperty("java.library.path", "");
+        if (libpath.length() > 0) {
+            paths.addAll(Arrays.asList(libpath.split(File.pathSeparator)));
+        }
+        ArrayList<URL> urls = new ArrayList<URL>(styles.length * (1 + paths.size()));
         for (int i = 0; cls != null && i < styles.length; i++) {
             // ... then find it from in our resources ...
             URL u = cls.getResource(subdir + styles[i]);
             if (u != null) {
-                urls[k++] = u;
+                urls.add(u);
             }
         }
-        // ... and in case of bad resources, search the paths as well.
+        // ... and in case of bad resources search the paths last, or first on user request.
+        int k = pathsFirst ? 0 : urls.size();
         for (int i = 0; paths.size() > 0 && i < styles.length; i++) {
             for (String path : paths) {
                 File file = new File(path, styles[i]);
                 if (file.exists()) {
                     try {
-                        urls[k++] = file.toURI().toURL();
+                        urls.add(k++, file.toURI().toURL());
                     } catch (IOException ex) {
                         throw new RuntimeException(ex);
                     }
                 }
             }
         }
-        URL[] newurls = new URL[k];
-        System.arraycopy(urls, 0, newurls, 0, k);
-        return newurls;
+        return urls.toArray(new URL[urls.size()]);
     }
 
     /**
@@ -561,7 +584,7 @@ public class Loader {
      * @param urls the URLs to try loading the library from
      * @param libnameversion the name of the library + "@" + optional version tag
      * @return the full path of the file loaded, or the library name if unknown
-     *         (but {@code if (!loadLibraries) { return null; }})
+     *         (but {@code if (!isLoadLibraries) { return null; }})
      * @throws UnsatisfiedLinkError on failure
      */
     public static String loadLibrary(URL[] urls, String libnameversion) {
@@ -580,32 +603,27 @@ public class Loader {
         try {
             for (URL url : urls) {
                 File file;
-                try {
-                    // ... if the URL is not already a file ...
-                    file = new File(url.toURI());
-                } catch (Exception e) {
-                    // ... then check if it has not already been extracted, and if not ...
-                    if (!(file = new File(getCacheDir() != null ? getCacheDir() : getTempDir(), new File(url.getPath()).getName())).exists()) {
-                        if (tempFile != null && tempFile.exists()) {
-                            tempFile.deleteOnExit();
-                        }
-                        // ... then extract it from our resources ...
-                        if (logger.isDebugEnabled()) {
-                            logger.debug("Extracting " + url);
-                        }
-                        if (getCacheDir() != null) {
-                            file = extractResource(url, getCacheDir(), null, null);
-                        } else {
-                            file = tempFile = extractResource(url, getTempDir(), null, null);
-                        }
-                    } else while (System.currentTimeMillis() - file.lastModified() < 1000) {
-                        // ... else wait until the file is at least 1 second old ...
-                        try {
-                            Thread.sleep(1000);
-                        } catch (InterruptedException ex) {
-                            // ... reset interrupt to be nice ...
-                            Thread.currentThread().interrupt();
-                        }
+                // ... then check if it has not already been extracted, and if not ...
+                if (!(file = new File(getCacheDir() != null ? getCacheDir() : getTempDir(), new File(url.getPath()).getName())).exists()) {
+                    if (tempFile != null && tempFile.exists()) {
+                        tempFile.deleteOnExit();
+                    }
+                    // ... then extract it from our resources ...
+                    if (logger.isDebugEnabled()) {
+                        logger.debug("Extracting " + url);
+                    }
+                    if (getCacheDir() != null) {
+                        file = extractResource(url, getCacheDir(), null, null);
+                    } else {
+                        file = tempFile = extractResource(url, getTempDir(), null, null);
+                    }
+                } else while (System.currentTimeMillis() - file.lastModified() < 1000) {
+                    // ... else wait until the file is at least 1 second old ...
+                    try {
+                        Thread.sleep(1000);
+                    } catch (InterruptedException ex) {
+                        // ... reset interrupt to be nice ...
+                        Thread.currentThread().interrupt();
                     }
                 }
                 if (file != null && file.exists()) {